fix: test_configs
