+++ tests/producer.rs
@@ -2,7 +2,7 @@ use crate::utils::admin::{create_admin_client, create_topic};
 use crate::utils::consumer::{create_subscribed_base_consumer, poll_x_times_for_messages};
 use crate::utils::containers::KafkaContext;
 use crate::utils::logging::init_test_logger;
-use crate::utils::producer::{create_producer, send_record};
+use crate::utils::producer::base_producer::create_producer;
 use crate::utils::rand::rand_test_topic;
 use rdkafka::producer::BaseRecord;
 use rdkafka::Message;
@@ -58,7 +58,8 @@ pub async fn test_basic_produce() {
     let record = BaseRecord::to(&test_topic_name) // destination topic
         .key(&[1, 2, 3, 4]) // message key
         .payload("content"); // message payload
-    let send_record_result = send_record(&base_producer, record).await;
+    let send_record_result =
+        crate::utils::producer::base_producer::send_record(&base_producer, record).await;
     if send_record_result.is_err() {
         panic!("could not send record: {}", send_record_result.unwrap_err());
     }
